+++ wgs-website/src/AppRoutes.tsx
@@ -3,6 +3,8 @@ import { Layout } from "./components/Layout";
 import { Home } from "./pages/Home";
 import { Docs } from "./pages/Docs";
 import { DocsSSHDockerSetup } from "./pages/DocsSSHDockerSetup";
+import { Blog } from "./pages/Blog";
+import { BlogWhyWorkgridStudio } from "./pages/BlogWhyWorkgridStudio";
 import { Changelog } from "./pages/Changelog";
 import { About } from "./pages/About";
 import { Contact } from "./pages/Contact";
@@ -18,6 +20,8 @@ export function AppRoutes() {
         <Route index element={<Home />} />
         <Route path="docs" element={<Docs />} />
         <Route path="docs/ssh-docker-setup" element={<DocsSSHDockerSetup />} />
+        <Route path="blog" element={<Blog />} />
+        <Route path="blog/why-workgrid-studio" element={<BlogWhyWorkgridStudio />} />
         <Route path="changelog" element={<Changelog />} />
         <Route path="about" element={<About />} />
         <Route path="contact" element={<Contact />} />